Improved confirmation window. Now appears only if fields have been edited
