+++ src/context/modal-context.tsx
@@ -8,6 +8,8 @@ import {
 } from 'react'
 import PopupDialog from '~/components/popup-dialog/PopupDialog'
 import { PaperProps } from '@mui/material/Paper'
+import { useTranslation } from 'react-i18next'
+import useConfirm from '~/hooks/use-confirm'
 
 interface Component {
   component: React.ReactElement
@@ -31,6 +33,8 @@ const ModalProvider: FC<ModalProviderProps> = ({ children }) => {
   const [modal, setModal] = useState<React.ReactElement | null>(null)
   const [paperProps, setPaperProps] = useState<PaperProps>({})
   const [timer, setTimer] = useState<NodeJS.Timeout | null>(null)
+  const { t } = useTranslation()
+  const { checkConfirmation } = useConfirm()
 
   const closeModal = useCallback(() => {
     setModal(null)
@@ -61,13 +65,23 @@ const ModalProvider: FC<ModalProviderProps> = ({ children }) => {
     [closeModal, openModal]
   )
 
+  const closeWithConfirmation = useCallback(async () => {
+    const confirmed = checkConfirmation({
+      title: t('titles.confirmTitle'),
+      message: t('questions.unsavedChanges')
+    })
+    if (await confirmed) {
+      closeModal()
+    }
+  }, [checkConfirmation])
+
   return (
     <ModalContext.Provider value={contextValue}>
       {children}
       {modal && (
         <PopupDialog
           closeModalAfterDelay={closeModalAfterDelay}
-          onClose={closeModal}
+          onClose={closeWithConfirmation}
           content={modal}
           paperProps={paperProps}
           timerId={timer}
